+++ openeo/internal/graph_building.py
@@ -22,7 +22,7 @@ from openeo.util import dict_no_none, load_json_resource
 class FlatGraphableMixin(metaclass=abc.ABCMeta):
     """
     Mixin for classes that can be exported/converted to
-    a "flat graph" representation of a process graph.
+    a "flat graph" representation of an openEO process graph.
     """
 
     @abc.abstractmethod
@@ -208,10 +208,14 @@ class PGNode(_FromNodeMixin, FlatGraphableMixin):
         return PGNodeGraphUnflattener.unflatten(flat_graph=flat_graph, parameters=parameters)
 
 
-def as_flat_graph(x: Union[dict, FlatGraphableMixin, Any]) -> Dict[str, dict]:
+def as_flat_graph(x: Union[dict, FlatGraphableMixin, Path, Any]) -> Dict[str, dict]:
     """
     Convert given object to a internal flat dict graph representation.
     """
+    # TODO: document or verify which process graph flavor this is:
+    #       including `{"process": {"process_graph": {nodes}}` ("process graph with metadata")
+    #       including `{"process_graph": {nodes}}` ("process graph")
+    #       or just the raw process graph nodes?
     if isinstance(x, dict):
         return x
     elif isinstance(x, FlatGraphableMixin):